use context to make current user accessible from anywhere in the application
