+++ client/src/context/AuthContext.js
@@ -11,7 +11,7 @@ import {
 
 import React, { useContext, useState, useEffect } from 'react';
 import { auth, db } from '../controller/Firebase';
-import { collection, addDoc } from 'firebase/firestore';
+import { doc, setDoc } from 'firebase/firestore';
 
 const AuthContext = React.createContext();
 
@@ -23,18 +23,29 @@ export function AuthProvider({ children }) {
   const [currentUser, setCurrentUser] = useState(true);
   const [loading, setLoading] = useState();
 
-  function signup(email, password, displayName) {
-    return createUserWithEmailAndPassword(auth, email, password).then((res) => {
+  const images = [
+    'https://i.imgur.com/NDFE7BQ.jpg',
+    'https://i.imgur.com/Ks2oou4.jpg',
+    'https://i.imgur.com/kLcZbQT.jpeg',
+    'https://i.imgur.com/ncnHn9I.jpg',
+    'https://pic.onlinewebfonts.com/svg/img_258083.png',
+  ];
+  
+  const photoURL = images[Math.random() * images.length];
+  
+  async function signup(email, password, displayName) {
+
+    return await createUserWithEmailAndPassword(auth, email, password).then((res) => {
       updateProfile(res.user, {
         displayName,
-        photoURL: 'https://pic.onlinewebfonts.com/svg/img_258083.png',
+        photoURL,
       });
       try {
-        const docRef = addDoc(collection(db, 'users', res.user.id), {}).then(
-          () => {
-            console.log('Document written with ID: ', docRef.id);
-          }
-        );
+        setDoc(doc(db, 'users', res.user.uid), {
+          displayName,
+          photoURL,
+          likedPosts: [],
+        });
       } catch (e) {
         console.error('Error adding document: ', e);
       }
